Split cli test file
