+++ cli/Cargo.toml
@@ -7,8 +7,7 @@ name = "parsec_cli"
 publish.workspace = true
 repository.workspace = true
 version.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+autotests = false
 
 [features]
 testenv = []
