+++ cli/tests/integration/list_devices.rs
@@ -1,0 +1,29 @@
+use assert_cmd::Command;
+
+use libparsec::{tmp_path, TmpPath};
+
+use super::{get_testenv_config, run_local_organization, set_env};
+use crate::utils::{GREEN, RESET, YELLOW};
+
+#[rstest::rstest]
+#[tokio::test]
+async fn list_devices(tmp_path: TmpPath) {
+    let _ = env_logger::builder().is_test(true).try_init();
+    let tmp_path_str = tmp_path.to_str().unwrap();
+    let config = get_testenv_config();
+    let (url, _, _) = run_local_organization(&tmp_path, None, config)
+        .await
+        .unwrap();
+    set_env(tmp_path_str, &url);
+
+    let path = tmp_path.join("config/parsec3/libparsec");
+    let path_str = path.to_string_lossy();
+
+    Command::cargo_bin("parsec_cli")
+        .unwrap()
+        .arg("list-devices")
+        .assert()
+        .stdout(predicates::str::contains(format!(
+            "Found {GREEN}4{RESET} device(s) in {YELLOW}{path_str}{RESET}:"
+        )));
+}
