+++ cli/tests/integration/status_organization.rs
@@ -1,0 +1,44 @@
+use assert_cmd::Command;
+
+use libparsec::{tmp_path, TmpPath};
+
+use super::{get_testenv_config, run_local_organization, set_env};
+use crate::testenv_utils::DEFAULT_ADMINISTRATION_TOKEN;
+
+#[rstest::rstest]
+#[tokio::test]
+async fn status_organization(tmp_path: TmpPath) {
+    let _ = env_logger::builder().is_test(true).try_init();
+    let tmp_path_str = tmp_path.to_str().unwrap();
+    let config = get_testenv_config();
+    let (url, _, org_id) = run_local_organization(&tmp_path, None, config)
+        .await
+        .unwrap();
+
+    set_env(tmp_path_str, &url);
+
+    let expect = format!(
+        "{:#}\n",
+        serde_json::json!({
+            "active_users_limit": null,
+            "is_bootstrapped": true,
+            "is_expired": false,
+            "minimum_archiving_period": 2592000,
+            "user_profile_outsider_allowed": true
+        })
+    );
+
+    Command::cargo_bin("parsec_cli")
+        .unwrap()
+        .args([
+            "status-organization",
+            "--organization-id",
+            org_id.as_ref(),
+            "--addr",
+            &url.to_string(),
+            "--token",
+            DEFAULT_ADMINISTRATION_TOKEN,
+        ])
+        .assert()
+        .stdout(expect);
+}
